\name{dualCh}
\alias{dualCh}

\title{A sample cellHTS object containing dual channel data}

\description{
 Archived \code{\linkS4class{cellHTS}} object corresponding to the first three 384-well plates 
 of a genome-wide RNAi screen for pathway activity 
 in \emph{Drosophila} cells}

\usage{##cellHTS object, see examples for details}

\format{\code{\linkS4class{cellHTS}} object}

\examples{
    data(dualCh)
}
\keyword{datasets}
