\name{KcViabSmall}
\alias{KcViabSmall}

\title{A sample cellHTS object - D. melanogaster genome-wide RNAi screen}

\description{
 Archived \code{\linkS4class{cellHTS}} object corresponding to the first three 384-well plates 
 of a genome-wide RNAi screen of cell viability 
 in \emph{Drosophila} Kc167 cells}

\usage{##cellHTS object, see examples for details}

\format{\code{\linkS4class{cellHTS}} object}

\references{
Boutros, M., Kiger, A.A., Armknecht,S., Kerr,K., Hild,M., Koch,B., Haas, S.A., Heidelberg Fly Array Consortium, Paro,R. and Perrimon, N. (2004) Genome-wide RNAi analysis of growth and viability in Drosophila cells, \emph{Science} \bold{303}:832--5.
}

\examples{
    data(KcViabSmall)
}
\keyword{datasets}
