\name{oldKcViabSmall}
\alias{oldKcViabSmall}

\title{A sample S3 class cellHTS object - D. melanogaster genome-wide RNAi screen}

\description{
 Archived S3 class \code{cellHTS} object corresponding to the first three 384-well plates 
 of a genome-wide RNAi screen of cell viability 
 in \emph{Drosophila} Kc167 cells. 
 This data set was assembled and stored using \code{cellHTS} package and is used to show how to convert an old S3 class cellHTS object to the new S4 data class associated with \code{cellHTS2} package.
}

\usage{##cellHTS object, see examples for details}

\format{
S3 class \code{cellHTS} object as detailed in package \code{cellHTS}.
}

\references{
Boutros, M., Kiger, A.A., Armknecht,S., Kerr,K., Hild,M., Koch,B., Haas, S.A., Heidelberg Fly Array Consortium, Paro,R. and Perrimon, N. (2004) Genome-wide RNAi analysis of growth and viability in Drosophila cells, \emph{Science} \bold{303}:832--5.
}

\examples{
    data(oldKcViabSmall)
}
\keyword{datasets}
