\name{getEnVisionRawData}
\alias{getEnVisionRawData}
\alias{getEnVisionCrosstalkCorrectedData}
\title{Read a plate file obtain from EnVision Plate Reader}
\description{

 Import functions to read a plate file obtained from EnVision Plate
 Reader.  These functions should be set as the import function of
 \code{\link[cellHTS2:readPlateList]{readPlateList}} through the
 argument \code{importFun} when reading plate result files obtained from
 EnVision plate reader.
 
 }
\usage{

getEnVisionRawData(f, p)
getEnVisionCrosstalkCorrectedData(f, p)

}
\arguments{
  
  \item{f}{the name of the result plate file to read.}
  
  \item{p}{capturing the additional path argument, which will actually
    be ignored. \code{f} is expected to be a fully resolved file path.}
  
}

\details{

  These functions should not be called directly. Instead, they should be
  set as the import function of \code{\link[cellHTS2]{readPlateList}}
  through the argument \code{importFun} when reading plate result files
  obtained from an EnVision plate reader.

}

\value{
  
  These functions return a list with two components. The first component
  should be a 'data.frame' with the following slots: \code{well} (a
  character vector with the well identifier in the plate) and \code{val}
  (the intensity values measured at each well).  The second component of
  this list should be a character vector containing a copy of the
  imported input data file (such as the output of
  \code{\link[base:readLines]{readLines}}). It should be suitable to be
  used as input for \code{\link[base:writeLines]{writeLines}}.
  %, since it will be used to reproduce the intensity files that are
  %linked in the HTML quality reports generated by
  %\code{\link[cellHTS:writeReport]{writeReport}}.

}


\author{Ligia Bras \email{ligia@ebi.ac.uk}}

\seealso{

\code{\link[cellHTS2]{readPlateList}}
}

\examples{

    plateFile <- system.file("EnVisionExample/XXX_1500.csv", package = "cellHTS2")
    onePlate <- getEnVisionRawData(plateFile)

   ## to get the cross talk corrected data:
   onePlate2 <- getEnVisionCrosstalkCorrectedData(plateFile)

}
\keyword{manip}
