\name{cellHTS-package}
\alias{cellHTS-package}
\docType{package}
\title{cellHTS Package Overview}
\description{
   cellHTS Package Overview
}
\details{
  Important data class: \code{\linkS4class{cellHTS}}. Full help on methods and
  associated functions is available from within class help
  pages.

  Data sets: \code{\link{KcViab}}, \code{\link{KcViabSmall}}.

  Introductory information is available from vignettes, type
  \code{openVignette()}.

  Class-specific methods: \code{\link{annotate}}, \code{\link{configure}}, 
  \code{\link{state}}, \code{\link{description}}, \code{\link{scores}}, \code{\link{scores<-}},
  \code{\link{rawdata}}, \code{\link{rawdata<-}}.

  Standard generic methods: \code{\link{show}}, \code{\link{initialize}}, 
     \code{\link{validObject}}.

  Other functions: \code{\link{readPlateData}}

  Full listing of documented articles is available in HTML view by
  typing \code{help.start()} and selecting \code{cellHTS} package from the
  \code{Packages} menu or via \code{library(help="cellHTS")}.

}

\author{Ligia Bras \email{ligia@ebi.ac.uk}}

\keyword{package}
